+++ nodes/knxUltimate-config.js
@@ -470,29 +470,19 @@ return msg;`, "helplink": "https://github.com/Supergiovane/node-red-contrib-knx-
 
                                 // 12/08/2020 Check wether is a learn (save) command or a activate (play) command.
                                 if (dest === input.topic || dest === input.topicSave) {
-
-                                    // 21/09/2020 Check wether the scene controller has been disabled
-                                    if (input.hasOwnProperty("disabled") && input.disabled === true) {
-                                        new Promise((resolve, reject) => {
-                                            input.setNodeStatus({ fill: "grey", shape: "dot", text: "Disabled", payload: "", GA: "", dpt: "", devicename: "" });
-                                            input.handleSend({ savescene: false, recallscene: false, savevalue: false, disabled: true });
-                                            resolve(true); // fulfilled
-                                        }).then(function () { }).catch(function () { });
-                                        return;
-                                    }
-
+                                   
                                     // Prepare the two messages to be evaluated directly into the Scene Controller node.
                                     new Promise((resolve, reject) => {
                                         if (dest === input.topic) {
                                             try {
                                                 let msgRecall = buildInputMessage({ _srcGA: src, _destGA: dest, _event: evt, _Rawvalue: rawValue, _inputDpt: input.dpt, _devicename: input.name ? input.name : "", _outputtopic: input.outputtopic, _oNode: null });
-                                                input.RecallScene(msgRecall.payload);
+                                                input.RecallScene(msgRecall.payload, false);
                                             } catch (error) { }
                                         } // 12/08/2020 Do NOT use "else", because both topics must be evaluated in case both recall and save have same group address.
                                         if (dest === input.topicSave) {
                                             try {
                                                 let msgSave = buildInputMessage({ _srcGA: src, _destGA: dest, _event: evt, _Rawvalue: rawValue, _inputDpt: input.dptSave, _devicename: input.name ? input.name : "", _outputtopic: dest, _oNode: null });
-                                                input.SaveScene(msgSave.payload);
+                                                input.SaveScene(msgSave.payload, false);
                                             } catch (error) { }
                                         }
                                         resolve(true); // fulfilled